handle code block and table differently
