+++ src/test/java/com/amazon/opendistroforelasticsearch/sql/doctest/core/dsl/Body.java
@@ -1,0 +1,32 @@
+/*
+ *   Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ *
+ *   Licensed under the Apache License, Version 2.0 (the "License").
+ *   You may not use this file except in compliance with the License.
+ *   A copy of the License is located at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   or in the "license" file accompanying this file. This file is distributed
+ *   on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+ *   express or implied. See the License for the specific language governing
+ *   permissions and limitations under the License.
+ */
+
+package com.amazon.opendistroforelasticsearch.sql.doctest.core.dsl;
+
+import java.util.Arrays;
+import java.util.stream.Collectors;
+
+public class Body {
+    private final String[] fieldValues;
+
+    Body(String... fieldValues) {
+        this.fieldValues = fieldValues;
+    }
+
+    @Override
+    public String toString() {
+        return Arrays.stream(fieldValues).collect(Collectors.joining(",", "{", "}"));
+    }
+}
