+++ src/test/java/com/amazon/opendistroforelasticsearch/sql/doctest/core/dsl/Formats.java
@@ -1,0 +1,37 @@
+/*
+ *   Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ *
+ *   Licensed under the Apache License, Version 2.0 (the "License").
+ *   You may not use this file except in compliance with the License.
+ *   A copy of the License is located at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   or in the "license" file accompanying this file. This file is distributed
+ *   on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+ *   express or implied. See the License for the specific language governing
+ *   permissions and limitations under the License.
+ */
+
+package com.amazon.opendistroforelasticsearch.sql.doctest.core.dsl;
+
+import com.amazon.opendistroforelasticsearch.sql.doctest.core.request.SqlRequestFormat;
+import com.amazon.opendistroforelasticsearch.sql.doctest.core.response.SqlResponseFormat;
+
+public class Formats {
+    private final SqlRequestFormat requestFormat;
+    private final SqlResponseFormat responseFormat;
+
+    Formats(SqlRequestFormat requestFormat, SqlResponseFormat responseFormat) {
+        this.requestFormat = requestFormat;
+        this.responseFormat = responseFormat;
+    }
+
+    SqlRequestFormat request() {
+        return requestFormat;
+    }
+
+    SqlResponseFormat response() {
+        return responseFormat;
+    }
+}
